ananya/requirements.txt
python-dotenv
requests
langchain
langchain-community
langchain-openai
openai
pypdf
beautifulsoup4
markdownify

